Move and refactor the ProcessMetrics class to the node‑commons module.
