+++ iotdb-core/node-commons/src/main/i18n/en/org/apache/iotdb/commons/i18n/ServiceMessages.java
@@ -136,6 +136,10 @@ public final class ServiceMessages {
   // ---- CpuUsageMetrics ----
   public static final String CPU_USAGE_UPDATE_TIME = "Time for update cpu usage is {} ns";
 
+  // ---- ProcessMetrics ----
+  public static final String FAILED_GET_PROCESS_RESIDENT_MEMORY =
+      "Failed to get process resident memory for pid {}";
+
   private ServiceMessages() {}
 
   public static final String UNKNOWN_SERVICE_TYPE = "Unknown ServiceType: ";
